requirements.txt
behave==1.2.5
boto==2.38.0
chardet==2.3.0
cryptography==1.2.3
enum34==1.1.2
greenlet==0.4.9
idna==2.0
ipaddress==1.0.16
lockfile==0.12.2
mysqlclient==1.3.7
ndg-httpsclient==0.4.0
parse==1.6.6
parse-type==0.3.4
pyasn1==0.1.9
pycrypto==2.6.1
PyHamcrest==1.9.0
pyOpenSSL==0.15.1
python-cloudfiles==1.7.10
pyxdg==0.25
requests==2.9.1
six==1.10.0
ssh==1.8.0
urllib3==1.13.1
websocket-client==0.37.0